requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
absl-py==0.12.0
    # via
    #   tensorboard
    #   tensorflow
astunparse==1.6.3
    # via tensorflow
cachetools==4.2.1
    # via google-auth
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
cycler==0.10.0
    # via matplotlib
datetime==4.3
    # via -r requirements.in
flatbuffers==1.12
    # via tensorflow
gast==0.3.3
    # via tensorflow
google-auth-oauthlib==0.4.4
    # via tensorboard
google-auth==1.29.0
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-pasta==0.2.0
    # via tensorflow
grpcio==1.32.0
    # via
    #   tensorboard
    #   tensorflow
h5py==2.10.0
    # via tensorflow
idna==2.10
    # via requests
joblib==1.0.1
    # via scikit-learn
keras-preprocessing==1.1.2
    # via tensorflow
kiwisolver==1.3.1
    # via matplotlib
lxml==4.6.3
    # via pandas-datareader
markdown==3.3.4
    # via tensorboard
matplotlib==3.4.1
    # via -r requirements.in
numpy==1.19.5
    # via
    #   -r requirements.in
    #   h5py
    #   keras-preprocessing
    #   matplotlib
    #   opt-einsum
    #   pandas
    #   scikit-learn
    #   scipy
    #   tensorboard
    #   tensorflow
oauthlib==3.1.0
    # via requests-oauthlib
opt-einsum==3.3.0
    # via tensorflow
pandas-datareader==0.9.0
    # via -r requirements.in
pandas==1.2.4
    # via
    #   -r requirements.in
    #   pandas-datareader
pillow==8.2.0
    # via matplotlib
protobuf==3.15.8
    # via
    #   tensorboard
    #   tensorflow
pyasn1-modules==0.2.8
    # via google-auth
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyparsing==2.4.7
    # via matplotlib
python-dateutil==2.8.1
    # via
    #   matplotlib
    #   pandas
pytz==2021.1
    # via
    #   datetime
    #   pandas
requests-oauthlib==1.3.0
    # via google-auth-oauthlib
requests==2.25.1
    # via
    #   pandas-datareader
    #   requests-oauthlib
    #   tensorboard
rsa==4.7.2
    # via google-auth
scikit-learn==0.24.1
    # via -r requirements.in
scipy==1.6.2
    # via scikit-learn
six==1.15.0
    # via
    #   absl-py
    #   astunparse
    #   cycler
    #   google-auth
    #   google-pasta
    #   grpcio
    #   h5py
    #   keras-preprocessing
    #   protobuf
    #   python-dateutil
    #   tensorboard
    #   tensorflow
tensorboard-plugin-wit==1.8.0
    # via tensorboard
tensorboard==2.4.1
    # via tensorflow
tensorflow-estimator==2.4.0
    # via tensorflow
tensorflow==2.4.1
    # via -r requirements.in
termcolor==1.1.0
    # via tensorflow
threadpoolctl==2.1.0
    # via scikit-learn
typing-extensions==3.7.4.3
    # via tensorflow
urllib3==1.26.4
    # via requests
werkzeug==1.0.1
    # via tensorboard
wheel==0.36.2
    # via
    #   astunparse
    #   tensorboard
    #   tensorflow
wrapt==1.12.1
    # via tensorflow
zope.interface==5.4.0
    # via datetime

# The following packages are considered to be unsafe in a requirements file:
# setuptools
